Migrate to use Batch2 (now named Batch) instead of Batch. This also splits the single_batch_run script and notebook to no longer include projects and auth. These each now have their own demo notebook.
