+++ 2017-10/ClientAPIs/tests/batch_runner_test.py
@@ -3,7 +3,7 @@ import sys
 sys.path.append('..')
 
 from adam import Service
-from adam import Batch2
+from adam import Batch
 from adam import PropagationParams
 from adam import OpmParams
 from adam import BatchRunManager
@@ -62,7 +62,7 @@ class BatchRunnerTest(unittest.TestCase):
             'object_id': 'TestObjId',
         })
         
-        return Batch2(propagation_params, opm_params)
+        return Batch(propagation_params, opm_params)
         
     def test_small_batch(self):
         num_batches = 3
